Move validation functions to util.ts file
